Moved shared code to shared folder; removed unused utils
