+++ client-v2/src/actions/ExternalArticles.js
@@ -1,23 +1,6 @@
 // @flow
 
 import type { Action } from 'types/Action';
-import type { ExternalArticle } from 'types/Shared';
-
-function externalArticlesReceived(articles: {
-  [string]: ExternalArticle
-}): Action {
-  return {
-    type: 'SHARED/EXTERNAL_ARTICLES_RECEIVED',
-    payload: articles
-  };
-}
-
-function requestCollectionArticles(): Action {
-  return {
-    type: 'SHARED/EXTERNAL_ARTICLES_GET_RECEIVE',
-    receivedAt: Date.now()
-  };
-}
 
 function errorReceivingCollectionArticles(error: string): Action {
   return {
@@ -28,8 +11,4 @@ function errorReceivingCollectionArticles(error: string): Action {
   };
 }
 
-export {
-  externalArticlesReceived,
-  requestCollectionArticles,
-  errorReceivingCollectionArticles
-};
+export { errorReceivingCollectionArticles };
