--- client-v2/src/components/Shared/GroupDisplay.js
@@ -1,35 +1,0 @@
-// @flow
-
-import * as React from 'react';
-import { connect } from 'react-redux';
-
-import { type State } from 'types/State';
-import { createArticlesInCollectionGroupSelector } from '../../selectors/shared';
-import Article from './Article';
-
-type ContainerProps = {
-  groupName: string,
-  collectionId: string, // eslint-disable-line react/no-unused-prop-types
-  stage: string // eslint-disable-line react/no-unused-prop-types
-};
-
-type Props = ContainerProps & {
-  articles: string[]
-};
-
-const GroupDisplay = ({ groupName, articles }: Props) => (
-  <div>
-    <h3>{groupName}</h3>
-    {articles.map(id => <Article key={id} id={id} />)}
-  </div>
-);
-
-const createMapStateToProps = () => {
-  const articlesSelector = createArticlesInCollectionGroupSelector();
-  // $FlowFixMe
-  return (state: State, props: ContainerProps): { articles: string[] } => ({
-    articles: articlesSelector(state, props)
-  });
-};
-
-export default connect(createMapStateToProps)(GroupDisplay);
