+++ client-v2/src/shared/actions/ExternalArticles.js
@@ -1,0 +1,12 @@
+import type { ExternalArticle } from '../types/ExternalArticle';
+
+function externalArticlesReceived(articles: {
+  [string]: ExternalArticle
+}): Action {
+  return {
+    type: 'SHARED/EXTERNAL_ARTICLES_RECEIVED',
+    payload: articles
+  };
+}
+
+export { externalArticlesReceived };
