+++ client-v2/src/shared/components/Collection.js
@@ -4,14 +4,18 @@ import React from 'react';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
 
-import { createCollectionSelector } from 'selectors/shared';
-import type { Collection } from 'types/Shared';
-import type { State } from 'types/State';
+import type { Collection } from '../types/Collection';
+import type { State } from '../types/State';
 import Article from './Article';
 import GroupDisplay from './GroupDisplay';
+import {
+  createCollectionSelector,
+  selectSharedState
+} from '../selectors/shared';
 
 type ContainerProps = {
-  id: string // eslint-disable-line react/no-unused-prop-types
+  id: string, // eslint-disable-line react/no-unused-prop-types
+  selectSharedState: (state: any) => State // eslint-disable-line react/no-unused-prop-types
 };
 
 type Props = ContainerProps & {
@@ -61,7 +65,12 @@ const createMapStateToProps = () => {
   const collectionSelector = createCollectionSelector();
   // $FlowFixMe
   return (state: State, props: ContainerProps): { collection: Collection } => ({
-    collection: collectionSelector(state, { collectionId: props.id })
+    collection: collectionSelector(
+      props.selectSharedState
+        ? props.selectSharedState(state)
+        : selectSharedState(state),
+      { collectionId: props.id }
+    )
   });
 };
