+++ client-v2/src/shared/reducers/externalArticlesReducer.js
