+++ client-v2/src/shared/selectors/__tests__/shared.spec.js
