+++ client-v2/src/shared/types/Article.js
@@ -1,0 +1,13 @@
+// @flow
+
+import type { ExternalArticle } from './ExternalArticle';
+
+type Article = ExternalArticle & {
+  uuid: string,
+  id: string,
+  frontPublicationDate: number,
+  publishedBy: string,
+  group?: number
+};
+
+export type { Article };
