+++ client-v2/src/shared/types/ExternalArticle.js
@@ -1,0 +1,8 @@
+// @flow
+
+type ExternalArticle = {
+  id: string,
+  headline: string
+};
+
+export type { ExternalArticle };
