+++ client-v2/src/shared/types/State.js
@@ -1,0 +1,5 @@
+import { type Reducers } from '../reducers/sharedReducer';
+import { type ExtractReturnType } from './Utility';
+
+export type State = $ObjMap<Reducers, ExtractReturnType>;
+export type RootState = { shared: State };
