+++ client-v2/src/shared/types/Utility.js
@@ -1,0 +1,4 @@
+// @flow
+
+export type ExtractReturnType = <V>((...args: any[]) => V) => V;
+export type $ReturnType<Fn> = $Call<<T>((...Iterable<any>) => T) => T, Fn>;
