+++ client-v2/src/shared/util/__tests__/shared.spec.js
