+++ client-v2/src/types/Action.js
@@ -1,13 +1,12 @@
 // @flow
 
-import { type Config } from './Config';
-
 /**
  * Need to add new types into here and union them with `Action` in order
  * for typing to work nicely in reducers
  */
 
-import type { Collection, ArticleFragment, ExternalArticle } from './Shared';
+import type { Action as SharedActions } from 'shared/types/Action';
+import { type Config } from './Config';
 
 type ActionError =
   | 'Could not fetch fronts config'
@@ -41,11 +40,6 @@ type PathUpdate = {
   path: string
 };
 
-type CollectionReceivedAction = {
-  type: 'SHARED/COLLECTION_RECEIVED',
-  payload: Collection
-};
-
 type RequestFrontCollectionAction = {
   type: 'FRONTS_COLLECTION_GET_RECEIVE',
   receivedAt: number
@@ -58,39 +52,15 @@ type ErrorInAction = {
   receivedAt: number
 };
 
-type ExternalArticlesReceived = {
-  type: 'SHARED/EXTERNAL_ARTICLES_RECEIVED',
-  payload: { [string]: ExternalArticle }
-};
-
-type RequestCollectionArticles = {
-  type: 'SHARED/EXTERNAL_ARTICLES_GET_RECEIVE',
-  receivedAt: number
-};
-
-type CollectionReceived = {
-  type: 'COLLECTION_RECEIVED',
-  payload: Collection
-};
-
-type ArticleFragmentsReceived = {
-  type: 'SHARED/ARTICLE_FRAGMENTS_RECEIVED',
-  payload: { [string]: ArticleFragment }
-};
-
 export type Action =
   | ConfigReceivedAction
   | FrontsConfigReceivedAction
   | RequestFrontsConfigAction
   | ClearError
   | PathUpdate
-  | CollectionReceivedAction
   | RequestFrontCollectionAction
   | ErrorInAction
-  | ExternalArticlesReceived
-  | RequestCollectionArticles
-  | CollectionReceived
-  | ArticleFragmentsReceived;
+  | SharedActions;
 
 export type ActionType = $ElementType<Action, 'type'>;
 export type { ActionError };
