+++ client-v2/src/types/State.js
@@ -1,6 +1,6 @@
 // @flow
 
 import { type Reducers } from 'reducers/rootReducer';
-import { type ExtractReturnType } from './Utility';
+import { type ExtractReturnType } from 'shared/types/Utility';
 
 export type State = $ObjMap<Reducers, ExtractReturnType>;
