--- client-v2/src/types/Utility.js
@@ -1,3 +1,0 @@
-// @flow
-
-export type ExtractReturnType = <V>((...args: any[]) => V) => V;
